MLCMS Ex1
Exercise sheet 1: Modeling of human crowds

In this exercise, you will learn how to model and simulate a human crowd with a simple cellular automaton.In general, cellular automata are a computational tool to model a complex system , i.e. a system that is built up of many interacting parts.